fixed 64bit tagging versions
--HG--
branch : 64bit

+++ src/memory/GenerationalCollector.cpp
@@ -30,7 +30,7 @@ GenerationalCollector::GenerationalCollector(Heap* heap) : GarbageCollector(heap
 VMOBJECT_PTR mark_object(VMOBJECT_PTR obj) {
 #ifdef USE_TAGGING
 	//don't process tagged objects
-	if ((int32_t)((void*)obj) & 1)
+	if ((size_t)((void*)obj) & 1)
 		return obj;
 #endif
     if (obj->GetGCField() & MASK_OBJECT_IS_MARKED)
@@ -43,7 +43,7 @@ VMOBJECT_PTR mark_object(VMOBJECT_PTR obj) {
 VMOBJECT_PTR copy_if_necessary(VMOBJECT_PTR obj) {
 #ifdef USE_TAGGING
   //don't process tagged objects
-  if ((int32_t)((void*)obj) & 0x1)
+  if ((size_t)((void*)obj) & 0x1)
     return obj;
 #endif
   size_t gcField = obj->GetGCField();
